+++ src/pytest_redis/plugin.py
@@ -32,6 +32,10 @@ _help_compress = "Turn on redis dump files compression."
 _help_rdbchecksum = "Whether to add checksum to the rdb files"
 _help_syslog = "Whether to enable logging to the system logger"
 _help_save = "Redis persistance frequency configuration - seconds keys"
+_help_decode = (
+    "Client: to decode response or not. "
+    "See redis.StrictRedis decode_reponse client parameter."
+)
 
 
 def pytest_addoption(parser):
@@ -90,6 +94,12 @@ def pytest_addoption(parser):
         name='redis_syslog',
         type='bool',
         help=_help_syslog
+    )
+    parser.addini(
+        name='redis_decode',
+        type='bool',
+        help=_help_decode,
+        default=False
     )
 
     parser.addoption(
@@ -158,6 +168,12 @@ def pytest_addoption(parser):
         action="store_true",
         dest='redis_syslog',
         help=_help_syslog
+    )
+    parser.addoption(
+        '--redis-client-decode',
+        action="store_true",
+        dest='redis_decode',
+        help=_help_decode
     )