+++ tests/test_redis.py
@@ -16,10 +16,10 @@ def test_redis(redisdb):
     redisdb.set('test2', 'test')
 
     test1 = redisdb.get('test1')
-    assert test1 == 'test'
+    assert test1 == b'test'
 
     test2 = redisdb.get('test2')
-    assert test2 == 'test'
+    assert test2 == b'test'
 
 
 redis_proc2 = factories.redis_proc(port=6381)
@@ -34,13 +34,13 @@ def test_second_redis(redisdb, redisdb2):
     redisdb2.set('test2', 'test_other')
 
     test1 = redisdb.get('test1')
-    assert test1 == 'test'
+    assert test1 == b'test'
 
     test2 = redisdb.get('test2')
-    assert test2 == 'test'
+    assert test2 == b'test'
 
-    assert redisdb2.get('test1') == 'test_other'
-    assert redisdb2.get('test2') == 'test_other'
+    assert redisdb2.get('test1') == b'test_other'
+    assert redisdb2.get('test2') == b'test_other'
 
 
 redis_proc_to_mock = factories.redis_proc(port=None)